Fix text encoding bug that was preventing ODF files on paths with characters outside the 7-bit ASCII range from being opened using the odt get text plugin. The actual bug was in FileUnzip .
It's actually not clear what the encoding of file names /inside/ the zip
directory must be as this is not documented in the current ODF spec. I'm
waiting for clarification on this, but in the mean time using local8Bit()
because it's equivalent to ascii() for the 127 bit ASCII range
(unless you use shift-jis, and LOTS more will break with that) but has at
least some chance of working outside that range.


git-svn-id: svn://scribus.info/Scribus/branches/Version13x@6382 11d20701-8431-0410-a711-e3c959e3b870